Reorganized Code
Put all engine source code in an engine folder.

Added the renderer and swap chain class.

The renderer class will be used for rendering logic and the swap chain class will be used to set up the swap chain.
