+++ source/engine/renderer.cpp
@@ -1,0 +1,5 @@
+#include "./engine/renderer.h"
+
+namespace JCAT {
+    
+}
