+++ source/engine/renderer.h
@@ -1,0 +1,13 @@
+#ifndef RENDERER_H
+#define RENDERER_H
+
+#include <iostream>
+
+namespace JCAT {
+    class Renderer {
+        public:
+        private:
+    };
+}
+
+#endif //RENDERER_H
