+++ source/engine/swapChain.h
@@ -1,0 +1,13 @@
+#ifndef SWAP_CHAIN_H
+#define SWAP_CHAIN_H
+
+#include <iostream>
+
+namespace JCAT {
+    class SwapChain {
+        public:
+        private:
+    };
+}
+
+#endif //SWAP_CHAIN_H
