--- source/graphicsPipeline.cpp
@@ -1,5 +1,0 @@
-#include "graphicsPipeline.h"
-
-namespace JCAT {
-    
-}
